Replace tabs with spaces
I know it is a personal preference but I very much dislike tabs in readme markdown, the spacing is completely off compared to the rest of the code examples. A tab takes up much more space than 4 spaces. It is a personal gripe, feel free to ignore this change if you don't care.